+++ CHANGELOG.md
@@ -7,6 +7,14 @@
 - Added support for `onPermissionRequest` event on iOS 15.0+
 - Updated `getMetaThemeColor` on iOS 15.0+
 
+## 5.4.2
+
+- Added `setActionButton` method to `ChromeSafariBrowser` class
+
+## 5.4.1+2
+
+- Fixed "Android ServiceWorkerControllerCompat.setServiceWorkerClient(null) makes Webivew Plugin Crashes" [#1151](https://github.com/pichillilorenzo/flutter_inappwebview/issues/1151)
+
 ## 5.4.1+1
 
 - Fixed Android default context menu over custom context menu on API Level 31+